adding auth to admin & seller and fixing filter by name

+++ controllers/admin.js
@@ -1,10 +1,15 @@
 const adminModel = require("../models/admin");
 const bcrypt = require("bcrypt");
+const { default: mongoose } = require("mongoose");
 
 const getAllAdmins = async (req, res, next) => {
   try {
-    const newAdmin = await adminModel.find();
-    res.status(200).json(newAdmin);
+    if (req.role === "admin") {
+      const newAdmin = await adminModel.find();
+      res.status(200).json(newAdmin);
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.json({ message: err.message });
   }
@@ -30,44 +35,60 @@ const loginAdmin = async (req, res, next) => {
   }
 };
 const AddnewAdmin = async (req, res, next) => {
-  const admin = req.body;
   try {
-    const savedAdmin = await adminModel.create(admin);
-    const token = await savedAdmin.generateAuthToken();
-    res.status(201).json({ savedAdmin, token });
+    if (req.role === "admin") {
+      const admin = req.body;
+      const savedAdmin = await adminModel.create(admin);
+      const token = await savedAdmin.generateAuthToken();
+      res.status(201).json({ savedAdmin, token });
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
 };
 
 const getAdminById = async (req, res, next) => {
-  const { id } = req.params;
   try {
-    const specificAdmin = await adminModel.findById(id);
-    res.status(200).json(specificAdmin);
+    if (req.role === "admin") {
+      const { id } = req.params;
+      const specificAdmin = await adminModel.findById(id);
+      res.status(200).json(specificAdmin);
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.json({ message: err.message });
   }
 };
 
 const updateAdminById = async (req, res) => {
-  const id = req.params.id;
-  const obj = req.body;
   try {
-    let updatedAdmin = await adminModel.findByIdAndUpdate(id, obj, {
-      new: true,
-    }); //new(options) if true, return the modified document rather than the original
-    res.json(updatedAdmin);
+    if (req.role === "admin" && req.admin._id.equals(req.params.id)) {
+      const id = req.params.id;
+      const obj = req.body;
+      const updatedAdmin = await adminModel.findByIdAndUpdate(id, obj, {
+        new: true,
+      }); //new(options) if true, return the modified document rather than the original
+      res.json(updatedAdmin);
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
 };
 
 const deleteAdminById = async (req, res) => {
-  const id = req.params.id;
   try {
-    let deletedAdmin = await adminModel.findByIdAndDelete(id);
-    res.json("Admin deleted successfully");
+    if (req.role === "admin" && req.admin._id.equals(req.params.id)) {
+      const id = req.params.id;
+      await adminModel.findByIdAndDelete(id);
+      res.json("Admin deleted successfully");
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
