+++ controllers/subDepartments.js
@@ -27,10 +27,14 @@ const getSubDepById = async (req, res, next) => {
 };
 //Add new subDepartment
 const AddnewSubDep = async (req, res, next) => {
-  const newSubDepartment = req.body;
   try {
-    const addededSubDep = await SubDepartmentModel.create(newSubDepartment);
-    res.status(201).json(addededSubDep);
+    if (req.role === "admin") {
+      const newSubDepartment = req.body;
+      const addededSubDep = await SubDepartmentModel.create(newSubDepartment);
+      res.status(201).json(addededSubDep);
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
@@ -38,13 +42,17 @@ const AddnewSubDep = async (req, res, next) => {
 
 //Update subDepartment
 const updateSubDepById = async (req, res) => {
-  const id = req.params.id;
-  const obj = req.body;
   try {
-    let updatedSubDep = await SubDepartmentModel.findByIdAndUpdate(id, obj, {
-      new: true,
-    });
-    res.json(updatedSubDep);
+    if (req.role === "admin") {
+      const id = req.params.id;
+      const obj = req.body;
+      let updatedSubDep = await SubDepartmentModel.findByIdAndUpdate(id, obj, {
+        new: true,
+      });
+      res.json(updatedSubDep);
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
@@ -52,10 +60,14 @@ const updateSubDepById = async (req, res) => {
 
 //Delete SubDepartment
 const deleteSubDepartment = async (req, res) => {
-  const id = req.params.id;
   try {
-    await SubDepartmentModel.findByIdAndDelete(id);
-    res.json("Sub-Department Deleted Successfully");
+    if (req.role === "admin") {
+      const id = req.params.id;
+      await SubDepartmentModel.findByIdAndDelete(id);
+      res.json("Sub-Department Deleted Successfully");
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
